fix small typo
The reason I think it is a typo and there must be an "e" is, that I
don't understand the sentence without the "e" but it would make
perfectly sense with it. It could be however, that the reason I don't
understand this is, that I'm not a native speaking person though.