fix an error in the documentation
I'm not 100% sure that this is an error, but as far as I understand
e.g. line 276 and line 289 in sklearn_genetic/algorithms.py, this
is rather the number of individuals than the number of hyperparameters
fit. Couldn't find out what an "invalid fitness" is, though. I guess
the number of individuals that are not already evaluated by fitness?