switch a sample that is a better advertising
this example gives reproducable a steady increase in max_fitness
(the in my opinion interesting value). Unfortunately it depends on
a weird 'tol' range (much too high). When a better range is used,
this example is no better than before. I tried to hint the user to
this fact inside the documentation.